+++ src/multipac_testbench/instruments/instrument.py
@@ -13,8 +13,8 @@ from matplotlib.container import StemContainer
 from matplotlib.lines import Line2D
 from multipac_testbench.multipactor_test.reduction_info import ReductionInfo
 from multipac_testbench.util.filtering import (
-    array_is_growing,
     clean_boolean_mask,
+    noisy_array_is_growing,
     remove_trailing_true,
 )
 from multipac_testbench.util.types import CALLBACK_T, POST_TREATER_T
@@ -571,8 +571,8 @@ class Instrument(ABC):
 
         The method performs three main operations:
 
-        #. It uses a sliding-window heuristic (*via* :func:`.array_is_growing`)
-           to detect growth.
+        #. It uses a sliding-window heuristic (*via*
+           :func:`.noisy_array_is_growing`) to detect growth.
         #. It removes short, isolated ``False`` segments, enforcing a minimum
            number of consecutive ``True`` values to be considered valid.
         #. It clears any trailing ``True`` values near the end of the array to
@@ -591,7 +591,8 @@ class Instrument(ABC):
         width :
             Width of the sample to determine increase.
         **kwargs :
-            Additional keyword arguments passed to :func:`.array_is_growing`.
+            Additional keyword arguments passed to
+            :func:`.noisy_array_is_growing`.
 
         Returns
         -------
@@ -600,7 +601,7 @@ class Instrument(ABC):
         Notes
         -----
         - The detection is influenced by the choice of parameters and the
-          behavior of :func:`.array_is_growing`.
+          behavior of :func:`.noisy_array_is_growing`.
         - Trailing regions and short noise-like fluctuations are filtered out.
 
         .. todo::
@@ -612,7 +613,7 @@ class Instrument(ABC):
 
         local_is_growing = True
         for i in range(n_points):
-            local_is_growing = array_is_growing(
+            local_is_growing = noisy_array_is_growing(
                 self.data,
                 i,
                 no_change_value=local_is_growing,
@@ -639,10 +640,7 @@ class Instrument(ABC):
 
         return growth_mask
 
-    def growth_array(
-        self,
-        **kwargs,
-    ) -> NDArray[np.float64]:
+    def growth_array(self, **kwargs) -> NDArray[np.float64]:
         """Identify regions where the signal is increasing ("growing").
 
         This method analyzes a signal to determine where it exhibits a growing
@@ -650,8 +648,6 @@ class Instrument(ABC):
         signal, where ``1.0`` indicates a region of growth and ``-1.0``
         otherwise. ``0.0`` means constant signal.
         *A priori*, will be useful for:
-
-        - :class:`.PowerSetpoint` to determine power cycles
 
         Notes
         -----
@@ -665,7 +661,8 @@ class Instrument(ABC):
             Value to put in growth mask when we did not manage to find whether
             measured signal increased or not.
         **kwargs :
-            Additional keyword arguments passed to :func:`.array_is_growing`.
+            Additional keyword arguments passed to
+            :func:`.noisy_array_is_growing`.
 
         Returns
         -------
@@ -675,7 +672,7 @@ class Instrument(ABC):
         bool_to_float = {True: 1.0, False: -1.0, None: 0.0}
         is_growing = [
             bool_to_float[
-                array_is_growing(
+                noisy_array_is_growing(
                     self.data,
                     i,
                     width=2,